Move docker stream parser to ./pkg/logs/parser
